requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph
openai
tavily-python
selenium
beautifulsoup4
python-dotenv
chromadb